extract helper to utils, fix forms, fix required
