+++ components/Register/RegisterPage1.tsx
@@ -12,89 +12,7 @@ import {
   InputText,
   InputContainer,
 } from '@/styles/register.styles';
-
-/**
- * auto-format inputted phone number
- * adapted from https://tomduffytech.com/how-to-format-phone-number-in-react/
- * @param value inputted phone number
- * @returns phone number in the form of (xxx) xxx-xxxx
- */
-const formatPhoneNumber = (value: string) => {
-  // if no input, return
-  if (!value) return value;
-
-  // phone number only has numbers
-  const phoneNumber: string = value.replace(/[^\d]/g, '');
-
-  // length of phone number
-  const phoneNumberLength: number = phoneNumber.length;
-
-  // auto-format based on length of numbers inputted
-  if (phoneNumberLength < 4) return phoneNumber;
-
-  if (phoneNumberLength < 7) {
-    return `(${phoneNumber.slice(0, 3)}) ${phoneNumber.slice(3)}`;
-  }
-
-  return `(${phoneNumber.slice(0, 3)}) ${phoneNumber.slice(
-    3,
-    6
-  )}-${phoneNumber.slice(6, 10)}`;
-};
-
-/**
- * auto-format inputted birthday
- * @param value inputted birthday
- * @returns birthday in the form of MM-DD-YYYY
- */
-const formatBirthday = (value: string) => {
-  if (!value) return value;
-
-  const birthday: string = value.replace(/[^\d]/g, '');
-
-  const birthdayLength: number = birthday.length;
-
-  if (birthdayLength < 3) return birthday;
-
-  if (birthdayLength < 5) {
-    return `${birthday.slice(0, 2)}-${birthday.slice(2)}`;
-  }
-
-  return `${birthday.slice(0, 2)}-${birthday.slice(2, 4)}-${birthday.slice(
-    4,
-    8
-  )}`;
-};
-
-/**
- * validate inputted birthday in the format
- * adapted from https://bobbyhadz.com/blog/javascript-check-if-date-is-valid#validate-a-date-formatted-as-ddmmyyyy-in-javascript
- * @param dateStr inputted birthday MM-DD-YYYY
- * @returns true if input is a valid birthday
- */
-const dateIsValid = (dateStr: string) => {
-  const regex = /^\d{2}\-\d{2}\-\d{4}$/;
-
-  if (dateStr.match(regex) === null) {
-    return false;
-  }
-
-  // only changed the order of destructuring assignment
-  const [month, day, year] = dateStr.split('-');
-
-  // format Date string as `yyyy-mm-dd`
-  const isoFormattedStr = `${year}-${month}-${day}`;
-
-  const date = new Date(isoFormattedStr);
-
-  const timestamp = date.getTime();
-
-  if (typeof timestamp !== 'number' || Number.isNaN(timestamp)) {
-    return false;
-  }
-
-  return date.toISOString().startsWith(isoFormattedStr);
-};
+import { dateIsValid, formatBirthday, formatPhoneNumber } from '@/utils/utils';
 
 const RegisterPage1 = ({
   formFunctions: {
@@ -163,9 +81,7 @@ const RegisterPage1 = ({
 
           <InputContainer>
             <InputText
-              {...register('birthday', {
-                required: true,
-              })}
+              {...register('birthday', { required: true })}
               placeholder="Date of birth (MM-DD-YYYY)"
               value={birthdayValue}
               width="100%"
